feat(front): getting frnot ready for diff
