+++ targets/frontend/src/components/themes/MapModal.js
@@ -1,3 +1,5 @@
+/** @jsxImportSource theme-ui */
+
 import { useState } from "react";
 import { IoMdMap } from "react-icons/io";
 import { Button } from "src/components/button";
