+++ targets/frontend/src/pages/glossary/index.js
@@ -1,3 +1,5 @@
+/** @jsxImportSource theme-ui */
+
 import Link from "next/link";
 import { useEffect, useMemo, useState } from "react";
 import { IoMdAdd, IoMdCloseCircleOutline } from "react-icons/io";
